fixedspointed out by sonar

+++ src/TransformOperator.cpp
@@ -124,7 +124,7 @@ TransformOperator::~TransformOperator()
 
 void TransformOperator::swap(int& x, int& y)
 {
-    float temp = x;
+    int temp = x;
     x = y; y = temp;
 }
 const Ogre::ColourValue& TransformOperator::getSelectionBoxColour() const
@@ -300,8 +300,8 @@ Ogre::Ray TransformOperator::rayFromScreenPoint(const QPoint& pos)
 {
     if(m_pActiveWidget)
     {
-        int width = m_pActiveWidget->getViewport()->getActualWidth() / 2.0f;
-        int height = m_pActiveWidget->getViewport()->getActualHeight() / 2.0f;
+        int width = m_pActiveWidget->getViewport()->getActualWidth() / 2;
+        int height = m_pActiveWidget->getViewport()->getActualHeight() / 2;
 
         Ogre::Real x = (Ogre::Real)(pos.x()) / (Ogre::Real)width;
         Ogre::Real y = (Ogre::Real)(pos.y()) / (Ogre::Real)height;
@@ -428,8 +428,8 @@ void TransformOperator::mouseMoveEvent(QMouseEvent *e)
     {
         if(m_pSelectionBox->isVisible() && m_pActiveWidget)
         {
-            int width = m_pActiveWidget->getViewport()->getActualWidth() / 2.0f;
-            int height = m_pActiveWidget->getViewport()->getActualHeight() / 2.0f;
+            int width = m_pActiveWidget->getViewport()->getActualWidth() / 2;
+            int height = m_pActiveWidget->getViewport()->getActualHeight() / 2;
 
             float xStart  = (float)(mScreenStart.x())/(float)width*2.0f-1.0f;
             float xStop   = (float)(e->pos().x())/(float)width*2.0f-1.0f;
